--- a/Makefile
+++ b/Makefile
@@ -35,6 +35,10 @@ acido-search: acido/minhashes.db acido/acido-chunk1.fa.gz.sig
 acido-frontier-search: acido/minhashes.db acido/acido-chunk1.fa.gz.sig
 	python -m search.frontier_search acido/acido-chunk1.fa.gz.sig acido 0.1
 
+acido-frontier-search-optimized: acido/minhashes.db acido/acido-chunk1.fa.gz.sig
+	python -m search.frontier_search acido/acido-chunk1.fa.gz.sig acido 0.1  --purgatory
+
+
 ### 
 
 15genome-clean: