--- a/search/memoize.py
+++ b/search/memoize.py
@@ -6,6 +6,8 @@ def memoize(func):
     def memoized_func(*args, **kwargs):
         key = str(args) + str(kwargs)
         if key not in cache:
-            cache[key] = func(*args, **kwargs)
+            val = func(*args, **kwargs)
+            cache[key] = val
+            return val
         return cache[key]
     return memoized_func