@domoritz 2017-05-05
This pr also fixes some bugs in the frontier search.